replace_hardcoded pages
