+++ src/assets/css/style.scss
@@ -691,6 +691,19 @@ footer {
 
 .review_name {
   color: #00faeb;
+  font-family: "Coming Soon";
+  font-style: normal;
+  font-weight: 400;
+  font-size: 28px;
+  line-height: 34px;
+}
+
+.review_text {
+  font-family: "Coming Soon";
+  font-style: normal;
+  font-weight: 400;
+  font-size: 20px;
+  line-height: 24px;
 }
 
 /*++________map */
