+++ src/pages/Artists.js
@@ -1,13 +1,10 @@
 import React from "react";
 import { Parallax } from "react-parallax";
-// import { AnimationOnScroll } from "react-animation-on-scroll";
 import "atropos/css";
-import Atropos from "atropos/react";
-import LazyLoad from "react-lazy-load";
-import { Link } from "react-router-dom";
 import ScrollToTop from "../assets/components/common/ScrollToTop";
 import bg_artist from "./../assets/img/bg_artists_page.png";
 import artistsData from "../data/artists.json";
+import ArtistInfo from "../assets/components/ArtistInfo";
 
 function Artists() {
   return (
@@ -15,40 +12,9 @@ function Artists() {
       <main>
         <section className="artists_page">
           <div className="container">
-            {artistsData.map((artist, index) => (
-              <div
-                key={index}
-                className={`artist_page_item ${
-                  index % 2 === 1 ? "reverse" : ""
-                }`}
-              >
-                  <div className="artist_page_info">
-                    <Atropos shadow={false} highlight={false} rotateTouch={false}>
-                      <h2 className="title_h2">{artist.name}</h2>
-                    </Atropos>
-                    <p className="text1">{artist.description1}</p>
-                    <p className="text2">{artist.description2}</p>
-                    <Link to={artist.link}>
-                      <button type="button" className="btn btn_see btn_artist">
-                        See works
-                      </button>
-                    </Link>
-                  </div>
-                
-                  <div className="artist_photo">
-                    <Atropos shadow={false} highlight={false} rotateTouch={false}>
-                      <LazyLoad height={648} offset={100}>
-                        <img  src={`./assets/img/${artist.image}`} alt={`${artist.name}_master`} />
-                      </LazyLoad>
-                    </Atropos>
-                  </div>
-                <Link to={`/artists/${artist.link}`}>
-                  <button type="button" className="btn btn_see btn_artist_resp">
-                    See works
-                  </button>
-                </Link>
-              </div>
-            ))}
+          {artistsData.map((artist, index) => (
+  <ArtistInfo key={index} artist={{ ...artist, index }} />
+))}
           </div>
         </section>
       </main>
